fix(ci): resolve file size limit violations
Split persona runtime prompt/event formatting into a dedicated mixin module to keep __init__.py under 500 lines, and tighten changelog prose to stay within the 3000-word limit.
