+++ FILEMAP.md
@@ -1,14 +1,14 @@
 # Repository File Map
 
 > Auto-generated by `scripts/generate_filemap.py` — 2026-04-23  
-> **338** tracked files
+> **339** tracked files
 
 ## Summary
 
 | Top-level directory | Files |
 |---------------------|-------|
 | `docs/` | 75 |
-| `agents/` | 68 |
+| `agents/` | 69 |
 | `internal/` | 68 |
 | `tests/` | 41 |
 | `(root)` | 20 |
@@ -26,7 +26,7 @@
 
 | Extension | Count |
 |-----------|-------|
-| `.py` | 121 |
+| `.py` | 122 |
 | `.md` | 91 |
 | `.go` | 64 |
 | `.yaml` | 23 |
@@ -126,6 +126,7 @@
 │   │   ├── action_loop.py
 │   │   ├── memory_budget.py
 │   │   ├── memory_context.py
+│   │   ├── prompt_assembly.py
 │   │   └── state_persistence.py
 │   ├── persona_types.py
 │   ├── pyproject.toml
